backend/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
supabase==2.4.2
cloudinary==1.40.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
webauthn==2.1.0
emails==0.6
httpx==0.27.0
pydantic[email]==2.10.6
email-validator==2.2.0
